[low-power] update as review suggests
