+++ examples/common-switches.mk
@@ -153,7 +153,7 @@ COMMONCFLAGS                   += -DOPENTHREAD_CONFIG_MAC_CSL_RECEIVER_ENABLE=1
 endif
 
 ifeq ($(S2S),1)
-COMMONCFLAGS                   += -DOPENTHREAD_CONFIG_MAC_SED_TO_SED_ENABLE=1
+COMMONCFLAGS                   += -DOPENTHREAD_CONFIG_MAC_SSED_TO_SSED_LINK_ENABLE=1
 endif
 
 ifeq ($(CSL_DEBUG),1)
